# Performance improvements for `recordingSpan` `SetAttributes` and `addOverCapAttrs`