--- a/sdk/trace/span.go
+++ b/sdk/trace/span.go
@@ -224,13 +224,13 @@ func (s *recordingSpan) SetStatus(code codes.Code, description string) {
 // attributes the span is configured to have, the last added attributes will
 // be dropped.
 func (s *recordingSpan) SetAttributes(attributes ...attribute.KeyValue) {
-	if s == nil {
+	if s == nil || len(attributes) == 0 {
 		return
 	}
 
 	s.mu.Lock()
 	defer s.mu.Unlock()
-	if !s.isRecording() || len(attributes) == 0 {
+	if !s.isRecording() {
 		return
 	}