Good day,

While working on https://github.com/open-telemetry/opentelemetry-go/pull/5858 I found some other possible improvements.

This PR:
- Adds an early return to `SetAttributes` when no attributes are provided. 
- Only increases `s.attributes` to guarantee that there is enough space for elements to be added.
- Fixes and issue where `truncateAttr` was not used when a attribute was being updated in `addOverCapAttrs`.

Thanks for reviewing and please let me know if any changes are needed.