@dmathieu 2024-10-02
Could you post benchstat results showing the improvements?