@boekkooi-impossiblecloud 2024-10-03
Hey @dmathieu 
Below are the benchmark results based on the last commit.
```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/sdk/trace
cpu: 11th Gen Intel(R) Core(TM) i5-11400H @ 2.70GHz
                                              │ 1727937489-attr-old.txt │       1727935323-attr-new.txt       │
                                              │         sec/op          │    sec/op     vs base               │
RecordingSpanSetAttributes/WithLimit/false-12              6.579µ ±  9%   6.279µ ±  9%       ~ (p=0.089 n=10)
RecordingSpanSetAttributes/WithLimit/true-12               14.27µ ± 10%   13.59µ ± 15%       ~ (p=0.247 n=10)
SpanEnd-12                                                 63.44n ±  0%   63.46n ±  0%       ~ (p=0.644 n=10)
TraceStart/with_a_simple_span-12                           728.0n ±  3%   729.9n ±  3%       ~ (p=0.868 n=10)
TraceStart/with_several_links-12                           881.5n ±  1%   875.2n ±  2%       ~ (p=0.105 n=10)
TraceStart/with_attributes-12                              916.5n ±  2%   917.4n ±  4%       ~ (p=0.796 n=10)
SpanProcessorOnEnd/batch:_10,_spans:_10-12                 151.9n ±  3%   147.3n ±  3%  -3.00% (p=0.037 n=10)
SpanProcessorOnEnd/batch:_10,_spans:_100-12                1.497µ ±  2%   1.429µ ±  1%  -4.48% (p=0.000 n=10)
SpanProcessorOnEnd/batch:_100,_spans:_10-12                148.5n ±  2%   146.3n ±  1%       ~ (p=0.085 n=10)
SpanProcessorOnEnd/batch:_100,_spans:_100-12               1.481µ ±  3%   1.455µ ±  2%       ~ (p=0.171 n=10)
SpanProcessorVerboseLogging-12                             9.376µ ±  1%   8.993µ ±  2%  -4.09% (p=0.000 n=10)
SpanLimits/AttributeValueLengthLimit-12                    7.221µ ±  4%   6.916µ ±  5%       ~ (p=0.089 n=10)
SpanLimits/AttributeCountLimit-12                          6.300µ ±  6%   6.160µ ±  4%       ~ (p=0.280 n=10)
SpanLimits/EventCountLimit-12                              5.921µ ±  4%   5.907µ ±  5%       ~ (p=0.579 n=10)
SpanLimits/LinkCountLimit-12                               6.073µ ±  6%   5.889µ ±  3%       ~ (p=0.218 n=10)
SpanLimits/AttributePerEventCountLimit-12                  6.455µ ±  3%   6.171µ ±  6%  -4.40% (p=0.001 n=10)
SpanLimits/AttributePerLinkCountLimit-12                   6.263µ ±  4%   6.547µ ±  5%       ~ (p=0.123 n=10)
SpanSetAttributesOverCapacity-12                           1.683µ ±  1%   1.639µ ±  1%  -2.64% (p=0.000 n=10)
StartEndSpan/AlwaysSample-12                               803.0n ±  4%   775.4n ±  9%       ~ (p=0.239 n=10)
StartEndSpan/NeverSample-12                                216.0n ±  2%   212.0n ±  2%       ~ (p=0.060 n=10)
SpanWithAttributes_4/AlwaysSample-12                       1.217µ ±  2%   1.176µ ±  6%       ~ (p=0.119 n=10)
SpanWithAttributes_4/NeverSample-12                        370.4n ±  5%   371.5n ±  5%       ~ (p=0.754 n=10)
SpanWithAttributes_8/AlwaysSample-12                       1.494µ ±  4%   1.429µ ±  5%  -4.35% (p=0.034 n=10)
SpanWithAttributes_8/NeverSample-12                        477.1n ±  3%   468.4n ±  6%       ~ (p=0.529 n=10)
SpanWithAttributes_all/AlwaysSample-12                     1.310µ ±  5%   1.261µ ±  3%  -3.82% (p=0.043 n=10)
SpanWithAttributes_all/NeverSample-12                      388.0n ±  5%   398.9n ±  8%       ~ (p=0.739 n=10)
SpanWithAttributes_all_2x/AlwaysSample-12                  1.664µ ±  2%   1.641µ ±  6%       ~ (p=0.306 n=10)
SpanWithAttributes_all_2x/NeverSample-12                   522.7n ±  7%   530.1n ±  5%       ~ (p=0.684 n=10)
SpanWithEvents_4/AlwaysSample-12                           1.411µ ±  4%   1.363µ ±  5%       ~ (p=0.184 n=10)
SpanWithEvents_4/NeverSample-12                            218.4n ±  3%   215.3n ±  2%       ~ (p=0.051 n=10)
SpanWithEvents_8/AlwaysSample-12                           1.997µ ±  5%   1.964µ ±  2%       ~ (p=0.424 n=10)
SpanWithEvents_8/NeverSample-12                            226.6n ±  3%   221.9n ±  2%       ~ (p=0.118 n=10)
SpanWithEvents_WithStackTrace/AlwaysSample-12             1016.5n ±  2%   966.5n ±  8%       ~ (p=0.143 n=10)
SpanWithEvents_WithStackTrace/NeverSample-12               242.0n ±  3%   243.0n ±  2%       ~ (p=0.868 n=10)
SpanWithEvents_WithTimestamp/AlwaysSample-12              1019.5n ±  3%   996.7n ±  3%       ~ (p=0.143 n=10)
SpanWithEvents_WithTimestamp/NeverSample-12                276.4n ±  3%   275.7n ±  2%       ~ (p=0.306 n=10)
TraceID_DotString-12                                       71.94n ±  3%   72.83n ±  4%       ~ (p=0.190 n=10)
SpanID_DotString-12                                        50.88n ±  3%   51.02n ±  1%       ~ (p=0.579 n=10)
geomean                                                    952.8n         936.0n        -1.76%

                                              │ 1727937489-attr-old.txt │        1727935323-attr-new.txt        │
                                              │          B/op           │     B/op      vs base                 │
RecordingSpanSetAttributes/WithLimit/false-12            7.062Ki ± 0%     7.062Ki ± 0%       ~ (p=1.000 n=10) ¹
RecordingSpanSetAttributes/WithLimit/true-12             7.564Ki ± 0%     7.564Ki ± 0%       ~ (p=0.233 n=10)
SpanEnd-12                                                 0.000 ± 0%       0.000 ± 0%       ~ (p=1.000 n=10) ¹
TraceStart/with_a_simple_span-12                           704.0 ± 0%       704.0 ± 0%       ~ (p=1.000 n=10) ¹
TraceStart/with_several_links-12                           880.0 ± 0%       880.0 ± 0%       ~ (p=1.000 n=10) ¹
TraceStart/with_attributes-12                              960.0 ± 0%       960.0 ± 0%       ~ (p=1.000 n=10) ¹
SpanProcessorOnEnd/batch:_10,_spans:_10-12                 0.000 ± 0%       0.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanProcessorOnEnd/batch:_10,_spans:_100-12                0.000 ± 0%       0.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanProcessorOnEnd/batch:_100,_spans:_10-12                0.000 ± 0%       0.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanProcessorOnEnd/batch:_100,_spans:_100-12               0.000 ± 0%       0.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanProcessorVerboseLogging-12                           10.94Ki ± 0%     10.94Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanLimits/AttributeValueLengthLimit-12                  10.76Ki ± 0%     10.76Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanLimits/AttributeCountLimit-12                        10.02Ki ± 0%     10.02Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanLimits/EventCountLimit-12                            9.594Ki ± 0%     9.594Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanLimits/LinkCountLimit-12                             9.203Ki ± 0%     9.203Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanLimits/AttributePerEventCountLimit-12                10.64Ki ± 0%     10.64Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanLimits/AttributePerLinkCountLimit-12                 10.64Ki ± 0%     10.64Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanSetAttributesOverCapacity-12                           768.0 ± 0%       768.0 ± 0%       ~ (p=1.000 n=10) ¹
StartEndSpan/AlwaysSample-12                               704.0 ± 0%       704.0 ± 0%       ~ (p=1.000 n=10) ¹
StartEndSpan/NeverSample-12                                144.0 ± 0%       144.0 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_4/AlwaysSample-12                     1.188Ki ± 0%     1.188Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_4/NeverSample-12                        400.0 ± 0%       400.0 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_8/AlwaysSample-12                     1.688Ki ± 0%     1.688Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_8/NeverSample-12                        656.0 ± 0%       656.0 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_all/AlwaysSample-12                   1.312Ki ± 0%     1.312Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_all/NeverSample-12                      464.0 ± 0%       464.0 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_all_2x/AlwaysSample-12                2.062Ki ± 0%     2.062Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_all_2x/NeverSample-12                   848.0 ± 0%       848.0 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_4/AlwaysSample-12                         1.188Ki ± 0%     1.188Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_4/NeverSample-12                            144.0 ± 0%       144.0 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_8/AlwaysSample-12                         1.812Ki ± 0%     1.812Ki ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_8/NeverSample-12                            144.0 ± 0%       144.0 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_WithStackTrace/AlwaysSample-12              800.0 ± 0%       800.0 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_WithStackTrace/NeverSample-12               160.0 ± 0%       160.0 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_WithTimestamp/AlwaysSample-12               824.0 ± 0%       824.0 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_WithTimestamp/NeverSample-12                184.0 ± 0%       184.0 ± 0%       ~ (p=1.000 n=10) ¹
TraceID_DotString-12                                       64.00 ± 0%       64.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanID_DotString-12                                        32.00 ± 0%       32.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                               ²                 +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                              │ 1727937489-attr-old.txt │       1727935323-attr-new.txt       │
                                              │        allocs/op        │ allocs/op   vs base                 │
RecordingSpanSetAttributes/WithLimit/false-12              15.00 ± 0%     15.00 ± 0%       ~ (p=1.000 n=10) ¹
RecordingSpanSetAttributes/WithLimit/true-12               20.00 ± 0%     20.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanEnd-12                                                 0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
TraceStart/with_a_simple_span-12                           14.00 ± 0%     14.00 ± 0%       ~ (p=1.000 n=10) ¹
TraceStart/with_several_links-12                           15.00 ± 0%     15.00 ± 0%       ~ (p=1.000 n=10) ¹
TraceStart/with_attributes-12                              16.00 ± 0%     16.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanProcessorOnEnd/batch:_10,_spans:_10-12                 0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanProcessorOnEnd/batch:_10,_spans:_100-12                0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanProcessorOnEnd/batch:_100,_spans:_10-12                0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanProcessorOnEnd/batch:_100,_spans:_100-12               0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanProcessorVerboseLogging-12                             155.0 ± 0%     155.0 ± 0%       ~ (p=1.000 n=10) ¹
SpanLimits/AttributeValueLengthLimit-12                    54.00 ± 0%     54.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanLimits/AttributeCountLimit-12                          50.00 ± 0%     50.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanLimits/EventCountLimit-12                              47.00 ± 0%     47.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanLimits/LinkCountLimit-12                               47.00 ± 0%     47.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanLimits/AttributePerEventCountLimit-12                  50.00 ± 0%     50.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanLimits/AttributePerLinkCountLimit-12                   50.00 ± 0%     50.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanSetAttributesOverCapacity-12                           15.00 ± 0%     15.00 ± 0%       ~ (p=1.000 n=10) ¹
StartEndSpan/AlwaysSample-12                               14.00 ± 0%     14.00 ± 0%       ~ (p=1.000 n=10) ¹
StartEndSpan/NeverSample-12                                2.000 ± 0%     2.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_4/AlwaysSample-12                       16.00 ± 0%     16.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_4/NeverSample-12                        3.000 ± 0%     3.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_8/AlwaysSample-12                       16.00 ± 0%     16.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_8/NeverSample-12                        3.000 ± 0%     3.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_all/AlwaysSample-12                     16.00 ± 0%     16.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_all/NeverSample-12                      3.000 ± 0%     3.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_all_2x/AlwaysSample-12                  16.00 ± 0%     16.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithAttributes_all_2x/NeverSample-12                   3.000 ± 0%     3.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_4/AlwaysSample-12                           17.00 ± 0%     17.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_4/NeverSample-12                            2.000 ± 0%     2.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_8/AlwaysSample-12                           18.00 ± 0%     18.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_8/NeverSample-12                            2.000 ± 0%     2.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_WithStackTrace/AlwaysSample-12              16.00 ± 0%     16.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_WithStackTrace/NeverSample-12               3.000 ± 0%     3.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_WithTimestamp/AlwaysSample-12               17.00 ± 0%     17.00 ± 0%       ~ (p=1.000 n=10) ¹
SpanWithEvents_WithTimestamp/NeverSample-12                4.000 ± 0%     4.000 ± 0%       ~ (p=1.000 n=10) ¹
TraceID_DotString-12                                       2.000 ± 0%     2.000 ± 0%       ~ (p=1.000 n=10) ¹
SpanID_DotString-12                                        2.000 ± 0%     2.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                               ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```